# docs/docs.go/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/sticker_packs": {
            "get": {
                "description": "Fetches all sticker packs for a given user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stickers"
                ],
                "summary": "Get all Sticker Packs for a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of sticker packs",
                        "name": "count",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of sticker packs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/forms.StickerPackOut"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/forms.ErrorForm"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/forms.ErrorForm"
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a new sticker pack",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stickers"
                ],
                "summary": "Add a new Sticker Pack",
                "parameters": [
                    {
                        "description": "Sticker Pack details",
                        "name": "sticker_pack",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.StickerPackForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created Sticker Pack",
                        "schema": {
                            "$ref": "#/definitions/forms.StickerPackOut"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/forms.ErrorForm"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/forms.ErrorForm"
                        }
                    }
                }
            }
        },
        "/api/sticker_packs/{id}": {
            "get": {
                "description": "Fetches a sticker pack by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stickers"
                ],
                "summary": "Get a specific Sticker Pack",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sticker Pack ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sticker Pack details",
                        "schema": {
                            "$ref": "#/definitions/forms.StickerPackOut"
                        }
                    },
                    "404": {
                        "description": "Sticker Pack not found",
                        "schema": {
                            "$ref": "#/definitions/forms.ErrorForm"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/forms.ErrorForm"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a sticker pack by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stickers"
                ],
                "summary": "Delete a sticker pack",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sticker Pack ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sticker Pack deleted"
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/forms.ErrorForm"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/forms.ErrorForm"
                        }
                    }
                }
            }
        },
        "/api/sticker_packs/{pack_name}": {
            "get": {
                "description": "Fetches a sticker pack by its name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stickers"
                ],
                "summary": "Get a Sticker Pack by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sticker Pack name",
                        "name": "pack_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sticker Pack details",
                        "schema": {
                            "$ref": "#/definitions/forms.StickerPackOut"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/forms.ErrorForm"
                        }
                    },
                    "404": {
                        "description": "Sticker Pack not found",
                        "schema": {
                            "$ref": "#/definitions/forms.ErrorForm"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/forms.ErrorForm"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "forms.ErrorForm": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "forms.StickerPackForm": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "stickers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "forms.StickerPackOut": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "stickers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}